Refactor to use DockerHub and new Recipe layout (#24)
TL;DR: This PR refactors out the images/ directory for the Prefect Agent & Worker images in favour of using the new
pangeo-forge-bakery-images DockerHub repository. It also adds the oisst_recipe.py Recipe which the Azure Bakery uses as this is now the new go-to for Recipe -> Flow conversion. To match the Azure Bakery, I've also switched from pipenv to poetry - I was getting fed up of flakey locking on pipenv